% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fantasypros.R
\name{get_fantasypros}
\alias{get_fantasypros}
\title{Get FantasyPro weekly projections}
\usage{
get_fantasypros(week)
}
\arguments{
\item{week}{week of the football season}
}
\value{
Projections using custom scoring
}
\description{
Get FantasyPro weekly projections
}
\examples{
get_fantasypros(week = 9)
}

