% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yahoo.R
\name{get_league_key}
\alias{get_league_key}
\title{Get league key for your fantasy league}
\usage{
get_league_key(league_id, token)
}
\arguments{
\item{league_id}{yahoo league id}

\item{token}{security token}
}
\description{
Get league key for your fantasy league
}

