% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yahoo.R
\name{get_rosters}
\alias{get_rosters}
\title{Get team rosters for your league}
\usage{
get_rosters(league_key, token)
}
\arguments{
\item{league_key}{league key for Yahoo fantasy football}

\item{token}{security token}
}
\description{
Get team rosters for your league
}

