% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yahoo.R
\name{get_token}
\alias{get_token}
\title{Get security token for Yahoo Fantasy API}
\usage{
get_token(token_file)
}
\arguments{
\item{token_file}{text file with key and secret}
}
\description{
Get security token for Yahoo Fantasy API
}

